backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.31
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
O365==2.0.27
requests==2.31.0
Werkzeug==3.0.1
email-validator==2.2.0
twilio==8.10.0
sendgrid==6.11.0
stripe==11.1.1
Pillow==11.0.0
openai==1.6.1
